refactor: Web-Member 구조 변경
