+++ tracky-core/src/main/java/kernel360/trackycore/core/infrastructure/repository/MemberRepository.java
@@ -1,11 +1,11 @@
-package kernel360.trackyweb.member.infrastructure.repository;
+package kernel360.trackycore.core.infrastructure.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import kernel360.trackyweb.member.domain.entity.MemberEntity;
+import kernel360.trackycore.core.domain.entity.MemberEntity;
 
 @Repository
 public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
